renames memory.h to hypre_memory.h to avoid compilation errors using CUDA+Clang including hypre's memory.h instead of the standard memory.h
[  0%] Building CUDA object CMakeFiles/HYPRE.dir/struct_mv/struct_matvec.c.o
In file included from <built-in>:1:
In file included from /u/acorriga/install/llvm-libstdcxx-22.1.0/lib/clang/22/include/__clang_cuda_runtime_wrapper.h:493:
In file included from /usr/local/install/cuda-12.4/include/curand_mtgp32_kernel.h:108:
/u/acorriga/hypre/src/utilities/memory.h:104:15: error: unknown type name 'HYPRE_MAYBE_UNUSED_FUNC'
  104 | static inline HYPRE_MAYBE_UNUSED_FUNC hypre_MemoryLocation
