+++ src/utilities/headers
@@ -47,7 +47,7 @@ cat printf.h                   >> $INTERNAL_HEADER
 cat error.h                    >> $INTERNAL_HEADER
 cat mpistubs.h                 >> $INTERNAL_HEADER
 cat smp.h                      >> $INTERNAL_HEADER
-cat memory.h                   >> $INTERNAL_HEADER
+cat hypre_memory.h             >> $INTERNAL_HEADER
 cat memory_tracker.h           >> $INTERNAL_HEADER
 cat omp_device.h               >> $INTERNAL_HEADER
 cat threading.h                >> $INTERNAL_HEADER
